requirements.txt
appdirs==1.4.4
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.2
cached-property==1.5.2
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.0.1
click==8.1.3
cryptography==38.0.4
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.3.0
docutils==0.19
et-xmlfile==1.1.0
exceptiongroup==1.1.0
exchangelib==4.9.0
fonttools==4.38.0
fpdf2==2.6.1
furl==2.1.3
graphviz==0.13.2
h11==0.14.0
htmldocx==0.0.6
hubspot-api-client==4.0.6
idna==3.4
importlib-metadata==4.13.0
isodate==0.6.1
java-access-bridge-wrapper==0.9.5
jsonpath-ng==1.5.3
jsonschema==4.17.3
lxml==4.9.2
mss==6.1.0
netsuitesdk==1.24.0
notifiers==1.3.3
ntlm-auth==1.5.0
O365==2.0.26
oauthlib==3.2.2
openpyxl==3.1.0
orderedmultidict==1.0.1
outcome==1.2.0
packaging==21.3
pdfminer.six==20221105
Pillow==9.4.0
platformdirs==3.0.0
ply==3.11
proxy-tools==0.1.0
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
pynput-robocorp-fork==5.0.0
pyotp==2.8.0
pyparsing==3.0.9
pypdf==3.4.0
pyperclip==1.8.2
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.21.1
python-xlib==0.33
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pywebview==3.6.3
PyYAML==5.4.1
requests==2.28.2
requests-file==1.5.1
requests-ntlm==1.1.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
robocorp-dialog==0.5.3
robotframework==5.0.1
robotframework-pythonlibcore==4.1.0
robotframework-requests==0.9.4
robotframework-seleniumlibrary==6.0.0
robotframework-seleniumtestability==2.0.0
rpaframework==21.1.0
rpaframework-core==10.2.1
rpaframework-dialogs==4.0.1
rpaframework-pdf==7.0.1
selenium==4.5.0
simple-salesforce==1.12.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stringcase==1.2.0
tenacity==8.2.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
tweepy==3.10.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
webdriver-manager==3.8.5
webdrivermanager==0.10.0
wrapt==1.14.1
wsproto==1.2.0
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
zeep==4.2.1
zipp==3.13.0
